Move new helpers to utf_utils.h
